Added keyboard handler for world movement and fixed some errors

+++ Entity.java
@@ -9,11 +9,11 @@ abstract class Entity{
 		active = true;
 	}
 	
-	public void setR(double r){this.r = r;}
-	public void setC(double c){this.c = c;}
+	public void setR(int r){this.r = r;}
+	public void setC(int c){this.c = c;}
 	public void setActive(boolean b){active = b;}
-	public double getR(){return r;}
-	public double getC(){return c;}
+	public int getR(){return r;}
+	public int getC(){return c;}
 	public boolean isActive(){return active;}
 	
 	public abstract void draw(Graphics2D g, int offsetX, int offsetY);
